Revert "test: Simplify zio layer setup for tests"
This reverts commit 51e8aa23a354f51301ed7b53b7350582781b44d8.
