+++ integration/src/test/scala/org/knora/webapi/core/LayersTestLive.scala
@@ -7,20 +7,15 @@ package org.knora.webapi.core
 
 import org.apache.pekko.actor.ActorSystem
 import zio.*
-import zio.ULayer
-import zio.ZLayer
 
 import org.knora.webapi.config.AppConfig
-import org.knora.webapi.config.InstrumentationServerConfig
+import org.knora.webapi.config.AppConfig.AppConfigurations
+import org.knora.webapi.messages.StringFormatter
 import org.knora.webapi.messages.util.*
-import org.knora.webapi.messages.util.search.QueryTraverser
-import org.knora.webapi.messages.util.search.gravsearch.transformers.OntologyInferencer
-import org.knora.webapi.messages.util.search.gravsearch.types.GravsearchTypeInspectionRunner
 import org.knora.webapi.messages.util.standoff.StandoffTagUtilV2
 import org.knora.webapi.messages.util.standoff.StandoffTagUtilV2Live
 import org.knora.webapi.responders.IriService
 import org.knora.webapi.responders.admin.*
-import org.knora.webapi.responders.admin.ListsResponder
 import org.knora.webapi.responders.v2.*
 import org.knora.webapi.responders.v2.ontology.CardinalityHandler
 import org.knora.webapi.routing.*
@@ -32,10 +27,12 @@ import org.knora.webapi.slice.admin.api.service.PermissionRestService
 import org.knora.webapi.slice.admin.api.service.ProjectRestService
 import org.knora.webapi.slice.admin.api.service.UserRestService
 import org.knora.webapi.slice.admin.domain.service.*
+import org.knora.webapi.slice.admin.domain.service.ProjectExportStorageService
 import org.knora.webapi.slice.common.ApiComplexV2JsonLdRequestParser
 import org.knora.webapi.slice.common.BaseModule
 import org.knora.webapi.slice.common.api.*
 import org.knora.webapi.slice.common.repo.service.PredicateObjectMapper
+import org.knora.webapi.slice.infrastructure.CacheManager
 import org.knora.webapi.slice.infrastructure.InfrastructureModule
 import org.knora.webapi.slice.infrastructure.OpenTelemetry
 import org.knora.webapi.slice.infrastructure.api.ManagementEndpoints
@@ -45,106 +42,64 @@ import org.knora.webapi.slice.lists.domain.ListsService
 import org.knora.webapi.slice.ontology.CoreModule
 import org.knora.webapi.slice.ontology.api.OntologyApiModule
 import org.knora.webapi.slice.ontology.domain.service.CardinalityService
-import org.knora.webapi.slice.ontology.domain.service.OntologyRepo
 import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 import org.knora.webapi.slice.resourceinfo.ResourceInfoLayers
+import org.knora.webapi.slice.resourceinfo.api.ResourceInfoEndpoints
 import org.knora.webapi.slice.resources.ResourcesModule
 import org.knora.webapi.slice.resources.api.ResourcesApiModule
-import org.knora.webapi.slice.resources.api.ResourcesApiRoutes
 import org.knora.webapi.slice.resources.repo.service.ResourcesRepo
 import org.knora.webapi.slice.resources.repo.service.ResourcesRepoLive
 import org.knora.webapi.slice.search.api.SearchApiRoutes
 import org.knora.webapi.slice.search.api.SearchEndpoints
+import org.knora.webapi.slice.security.ScopeResolver
 import org.knora.webapi.slice.security.SecurityModule
 import org.knora.webapi.slice.security.api.AuthenticationApiModule
+import org.knora.webapi.slice.security.api.AuthenticationApiRoutes
 import org.knora.webapi.slice.shacl.ShaclModule
 import org.knora.webapi.slice.shacl.api.ShaclApiModule
-import org.knora.webapi.slice.shacl.api.ShaclApiRoutes
 import org.knora.webapi.store.iiif.IIIFRequestMessageHandler
 import org.knora.webapi.store.iiif.IIIFRequestMessageHandlerLive
 import org.knora.webapi.store.iiif.api.SipiService
 import org.knora.webapi.store.iiif.impl.SipiServiceLive
 import org.knora.webapi.store.triplestore.upgrade.RepositoryUpdater
+import org.knora.webapi.testcontainers.DspIngestTestContainer
+import org.knora.webapi.testcontainers.FusekiTestContainer
+import org.knora.webapi.testcontainers.SharedVolumes
+import org.knora.webapi.testcontainers.SipiTestContainer
+import org.knora.webapi.testservices.TestClientService
+import org.knora.webapi.testservices.TestDspIngestClient
 
-object LayersLive {
+object LayersTestLive { self =>
 
-  /**
-   * The `Environment` that we require to exist at startup.
-   */
-  type DspEnvironmentLive =
-    // format: off
-    ActorSystem &
-    AdminApiEndpoints &
-    AdminModule.Provided &
-    ApiComplexV2JsonLdRequestParser &
-    ApiRoutes &
-    ApiV2Endpoints &
-    AppConfig.AppConfigurations &
-    AssetPermissionsResponder &
-    AuthenticationApiModule.Provided &
-    AuthorizationRestService &
-    BaseModule.Provided &
-    CardinalityHandler &
-    ConstructResponseUtilV2 &
-    CoreModule.Provided &
-    DefaultObjectAccessPermissionService &
-    GroupRestService &
-    HttpServer &
-    IIIFRequestMessageHandler &
-    InfrastructureModule.Provided &
-    InstrumentationServerConfig &
-    ListsApiModule.Provided &
-    ListsResponder &
-    ListsService &
-    MessageRelay &
-    OntologyApiModule.Provided &
-    OntologyInferencer &
-    OntologyResponderV2 &
-    PermissionRestService &
-    PermissionUtilADM &
-    PermissionsResponder &
-    ProjectExportService &
-    ProjectExportStorageService &
-    ProjectImportService &
-    ProjectRestService &
-    RepositoryUpdater &
-    ResourceUtilV2 &
-    ResourcesApiRoutes &
-    ResourcesResponderV2 &
-    ResourcesRepo &
-    SecurityModule.Provided &
-    SearchApiRoutes &
-    SearchResponderV2Module.Provided &
-    SecurityModule.Provided &
-    ShaclApiModule.Provided &
-    ShaclModule.Provided &
-    SipiService &
-    StandoffResponderV2 &
-    StandoffTagUtilV2 &
-    State &
-    UserRestService &
-    ValuesResponderV2
-    // format: on
+  type Environment =
+    LayersLive.DspEnvironmentLive & MessageRelayActorRef & FusekiTestContainer & TestClientService &
+      TestDspIngestClient & SipiTestContainer & DspIngestTestContainer & SharedVolumes.Volumes
 
   /**
-   * All effect layers needed to provide the `Environment`
+   * Provides a layer for integration tests which depend on Fuseki and Sipi as Testcontainers.
+   * @return a [[ULayer]] with the [[DefaultTestEnvironmentWithSipi]]
    */
-  val dspLayersLive: ULayer[DspEnvironmentLive] =
-    ZLayer.make[DspEnvironmentLive](
+  val layer: ULayer[self.Environment] =
+    ZLayer.make[self.Environment](
+      PekkoActorSystem.layer,
+      SipiServiceLive.layer,
+      TestContainerLayers.all,
+      TestDspIngestClient.layer,
+      TestClientService.layer,
+      /// common
       AdminApiModule.layer,
       AdminModule.layer,
       ApiComplexV2JsonLdRequestParser.layer,
       ApiRoutes.layer,
       ApiV2Endpoints.layer,
-      AppConfig.layer,
       AssetPermissionsResponder.layer,
       AuthenticationApiModule.layer,
       AuthorizationRestService.layer,
       BaseEndpoints.layer,
       BaseModule.layer,
       CardinalityHandler.layer,
-      ConstructResponseUtilV2.layer,
       CoreModule.layer,
+      ConstructResponseUtilV2.layer,
       DspIngestClientLive.layer,
       HandlerMapper.layer,
       HttpServer.layer,
@@ -157,11 +112,11 @@ object LayersLive {
       ListsService.layer,
       ManagementEndpoints.layer,
       ManagementRoutes.layer,
+      MessageRelayActorRef.layer,
       MessageRelayLive.layer,
       OntologyApiModule.layer,
       OntologyResponderV2.layer,
       OpenTelemetry.layer,
-      PekkoActorSystem.layer,
       PermissionUtilADMLive.layer,
       PermissionsResponder.layer,
       PredicateObjectMapper.layer,
@@ -181,12 +136,10 @@ object LayersLive {
       SecurityModule.layer,
       ShaclApiModule.layer,
       ShaclModule.layer,
-      SipiServiceLive.layer,
       StandoffResponderV2.layer,
       StandoffTagUtilV2Live.layer,
       State.layer,
       TapirToPekkoInterpreter.layer,
       ValuesResponderV2.layer,
-      // ZLayer.Debug.mermaid,
     )
 }
