+++ integration/src/test/scala/org/knora/webapi/testservices/TestClientService.scala
@@ -13,7 +13,9 @@ import org.apache.pekko.http.scaladsl.model.headers.HttpCredentials
 import org.apache.pekko.http.scaladsl.unmarshalling.Unmarshal
 import sttp.capabilities.zio.ZioStreams
 import sttp.client3
+import sttp.client3.*
 import sttp.client3.SttpBackend
+import sttp.client3.httpclient.zio.HttpClientZioBackend
 import zio.*
 import zio.json.*
 import zio.json.ast.Json
@@ -146,5 +148,13 @@ final case class TestClientService(
 }
 
 object TestClientService {
-  def layer = ZLayer.derive[TestClientService]
+  def layer: ZLayer[ActorSystem & AppConfig, Nothing, TestClientService] =
+    HttpClientZioBackend.layer().orDie >+>
+      ZLayer.scoped {
+        for {
+          sys    <- ZIO.service[ActorSystem]
+          config <- ZIO.service[AppConfig]
+          sttp   <- ZIO.service[SttpBackend[Task, ZioStreams]]
+        } yield TestClientService(config, sttp)(sys)
+      }
 }
